// _js/snippets/product-options/app/modules/selling-plan-dropdown/style.module.scss.d.ts
/* DO NOT EDIT: This file is auto-generated and will be overwritten. [build-fingerprint:codegen] */
export type Styles = {
  "sellingPlan": "sellingPlan",
  "sellingPlan__checkbox": "sellingPlan__checkbox",
  "sellingPlan__details": "sellingPlan__details",
  "sellingPlan__frequency": "sellingPlan__frequency",
  "sellingPlan__group": "sellingPlan__group",
  "sellingPlan__input": "sellingPlan__input",
  "sellingPlan__label": "sellingPlan__label",
  "sellingPlan__options": "sellingPlan__options",
  "sellingPlan__planPrice": "sellingPlan__planPrice",
  "sellingPlan__planPrice__before": "sellingPlan__planPrice__before",
  "sellingPlan__price": "sellingPlan__price",
  "sellingPlan__selector": "sellingPlan__selector",
  "sellingPlan__title": "sellingPlan__title"
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles